automation/requirements.txt
Appium-Python-Client==2.11.1
selenium==4.11.2